Move some tests to correct package.
