+++ core/src/test/java/nl/nn/adapterframework/senders/JavascriptSenderCallbackTest.java
@@ -1,4 +1,4 @@
-package nl.nn.adapterframework.extensions.javascript;
+package nl.nn.adapterframework.senders;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -20,11 +20,7 @@ import nl.nn.adapterframework.core.SenderException;
 import nl.nn.adapterframework.core.SenderResult;
 import nl.nn.adapterframework.core.TimeoutException;
 import nl.nn.adapterframework.parameters.Parameter.ParameterType;
-import nl.nn.adapterframework.senders.EchoSender;
-import nl.nn.adapterframework.senders.JavascriptSender;
 import nl.nn.adapterframework.senders.JavascriptSender.JavaScriptEngines;
-import nl.nn.adapterframework.senders.SenderTestBase;
-import nl.nn.adapterframework.senders.SenderWithParametersBase;
 import nl.nn.adapterframework.stream.Message;
 import nl.nn.adapterframework.testutil.ParameterBuilder;
